Update myScript.js
Added the final D3 code for the graph to scale right and implemented a tooltip using the mouseover event when hovering over the bars in the graph. The tooltip box is dynamically updated with the info extracted from the data-date and data-gdp attributes of every bar from the graph. The project is now final, passes all the tests in the "FCC Test Suite" and is submitted for review to the FCC.